# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for 0G Storage operations",
        "title": "0G Storage API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/download/{root_hash}": {
            "get": {
                "description": "Download a file using its root hash",
                "produces": [
                    "application/octet-stream"
                ],
                "summary": "Download a file from 0G Storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Root hash of the file",
                        "name": "root_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Upload a file to 0G Storage network",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a file to 0G Storage",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.UploadResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.UploadResponse": {
            "type": "object",
            "properties": {
                "root_hash": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                }
            }
        }
    }
}